Exported components and commonly used functions/constants in seperate files
